update: type definition
