+++ share/src/model.rs
@@ -1,1 +1,2 @@
-pub mod assemble;
+pub mod assemble_core;
+pub mod assemble_reqres;
